feat: Add preventEmpty to InputToggle in EditableTitle component
Add the preventEmpty prop to the InputToggle component in the EditableTitle component. This ensures that the input value cannot be empty. The change helps improve the user experience by preventing the submission of empty values.

fix: Update EditableDescription component

Create a new EditableDescription component that allows users to edit the project description. The component includes an InputToggle with a placeholder and a character limit of 64. The component also handles the onBlur event to update the project description via an API call. If the API call fails, an error notification is displayed.

fix: Update EditableTitle component

Create a new EditableTitle component that allows users to edit the project title. The component includes an InputToggle with a character limit of 64. The component also handles the onBlur event to update the project title via an API call. If the API call fails, an error notification is displayed.

refactor: Remove unused imports and variables

Remove unused imports and variables from the projectPageHeader component to improve code cleanliness and maintainability.
